Map class creation
Created the map class, it holds a matrix that represents the playable area. It also has methods for displaying the map and to move characters from one position to another.
On this map we will display other components like the player ship, enemies, eggs, etc. All of them will use MoveChar function.