Adding player input
Added the player input for the other arrow keys. Also added a method that checks for collisions so that the player can't leave the playable area.
Up until now all the commits implemented functionality that we will all have to use during development. We made these commits together so that we could have a starting point.