Finished egg + collision with player ship
The egg falls down on the screen and if it collides with the ship we erase the egg and kill the player. The egg is also erased once it leaves the screen.
I implemented the egg class with the move constructor because we need it to insert eggs into a container so that we can keep evidence of them.