TitleScreen
I have added a fully working titleScreen in which you can select either to leaderboard button or exit button . If ou have the exit button selected and press Enter ,the game will Quit. If you press Spacebar ,the game will start with wave 1 as usual. Escape button isn't working while in Title Screen because there is a button having this purpose but if you are in-game and press Escape ,the game will quit.