Optimising startup time
The game was taking too long to load all textures because we were making a lot of copies of them. I created a class that holds all the textures used in the game and that is responsible for loading them.
Each class should use the textures from this class instead of loading their own textures.
I currently changed a couple of classes to work with this new method: Player, Egg, Bullet, Present, Explosion, music, font and a couple of sprites.
In the next commit I will convert the rest of the classes and reorganize the project because the game loop is starting to look ugly.