Meat drop
I implemented the way a meat drops, it bounces off the side of the screen and the bottom part of it. When it bounces with the bottom of the screen the falling speed is decreased. The movement is calculated using acceleration.
Other than that I solved small errors in other classes.