Optimising startup even further
When starting up the game, each texture now is loaded on it's own thread, this reduced the loading time to almost half to what it was before. The reason as to why it's ONLY half is because the chicken sprite sheet is too large so it takes a long time to load it.