Input refactoring
I refactored the CheckInput method. Before, we checked the same key multiple times, which could have lead to some errors depending on user input. Now I reorganized all  of them, we check each key a single of time and do it's functionality based on the state of the game.
I also made it safer than it was before.